terminal-mcp-agent/requirements.txt
langgraph>=0.2.6
langchain>=0.3.19
langchain-google-genai
python-dotenv>=1.0.1
langgraph-sdk>=0.1.57
langgraph-cli
google-genai
